+++ src/java/businesslogic/Item.java
@@ -17,13 +17,13 @@ import java.util.Vector;
 public class Item {
     private int id;
     private String name;
-    private double price;
+    private float price;
     private int stock;
     /*static public Collection searchItems(String group, Connection con) {
         return ItemDB.searchItems(group, con);
     }*/
 
-    public Item(int id, String name, double price, int stock) {
+    public Item(int id, String name, float price, int stock) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -38,7 +38,7 @@ public class Item {
         return name;
     }
 
-    public double getPrice() {
+    public float getPrice() {
         return price;
     }
